Changed TouchLink interface to enable async operations.
TouchLink could only be implemented by a synchronous process: the
response of the server was returned immediately. For implementations
using message queues, reactive streams or other asynchronous protocols
this synchronous interface cannot be implemented, because the result is
not immediately available and other events should be able to fire in the
mean time.

The socket implementation does not use any async-ness.